Problem: `priv/react` clashes when made with c-r-a
Solution:

When we try to run

```
mkdir priv
cd priv
npx create-react-app react --template multi-nix
```

we get the following error:

```
Cannot create a project named "react" because a dependency with the same name exists.
Due to the way npm works, the following names are not allowed:

react
react-dom
react-scripts

Please choose a different project name.
```

To fix this, we have to change the default name of the UI sub-project. Thus, we change the nomenclature of `priv/react` to `priv/ui`.